% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiemcore.R
\name{kussakin}
\alias{kussakin}
\title{Use Kussakin's equation to calculate the range of a well funnel cone}
\usage{
kussakin(dh, kf, hm)
}
\arguments{
\item{dh}{the drawdown of the well in m}

\item{kf}{the hydraulic conductivity of the aquifer in m/s}

\item{hm}{the height of the water column above the quitard in m}
}
\value{
the range in m
}
\description{
\code{kussakin} needs the drawdown at the well, the
hydraulic potential of the undisturbed gw-surface as meters above aquitard
and the hydraulic conductiviry of the aquifer in which it is placed-
It will in return give you a fair estimate of the dimension of the ensuing
well funnel cone range.
}
\examples{
# first example
kussakin(dh = 2, kf = 10^-4, hm = 10)

# second example
aquifer_cond <- 2.5*10^-3 # m/s
aquifer_height <- 8 # m
drawdown_steps <- seq(0.2, 5, 0.2) # m

ranges <- vapply(
X = drawdown_steps,
FUN = kussakin,
FUN.VALUE = vector("double", length = 1),
kf = aquifer_cond, hm = aquifer_height)#' plot(ranges~drawdown_steps)

}
