% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiemcore.R
\name{sichardt}
\alias{sichardt}
\title{Use Sichardt's equation to calculate the range of a well funnel cone}
\usage{
sichardt(dh, kf)
}
\arguments{
\item{dh}{the drawdown of the well in m}

\item{kf}{the hydraulic conductivity of the aquifer in m/s}
}
\value{
the range in m
}
\description{
\code{sichardt} needs the drawdown at the well and the
hydraulic conductiviry of the aquifer in which it is placed and will give
you a fair estimate of the dimension of the ensuing well funnel cone range.
}
\examples{
# first example
sichardt(dh = 2, kf = 10^-4)

# second example
aquifer_cond <- 2.5*10^-3
drawdown_steps <- seq(0.2, 5, 0.2)

ranges <- vapply(
X = drawdown_steps,
FUN = sichardt,
FUN.VALUE = vector("double", length = 1),
kf = aquifer_cond)#' plot(ranges~drawdown_steps)

}
