% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiemcore.R
\name{thiem_coneshape}
\alias{thiem_coneshape}
\title{Use the Dupuit-Thiem equation to calculate the cone shape of a well}
\usage{
thiem_coneshape(
  x,
  Q,
  h0,
  r_well,
  kf,
  dh_search_min = 0.001,
  dh_search_max = h0
)
}
\arguments{
\item{x}{the distance in m from the well whre h is to be calculated}

\item{Q}{the pumping rate of the well in m³/s}

\item{h0}{the natural height of the water column above the aquitard (in absence of a well) in m}

\item{r_well}{the radius of the well}

\item{kf}{the hydraulic conductivity of the aquifer in m/s}

\item{dh_search_min}{the minimum possible drawdown of the well to be considered in m. Defaults to 0.001m == 1mm}

\item{dh_search_max}{the minimum possible drawdown of the well to be considered in m. Defaults to \code{h0}}
}
\value{
hydraulic potential \code{h} as m above aquitard in distance \code{x} to the well
}
\description{
\code{thiem_coneshape} calculates the hydraulic potential
\code{h} at a given distance \code{x} from a pumping well
}
\examples{
# first example
x_range <- seq(0.5, 80, 0.5)
h_at_x <- vapply(
x_range,
thiem_coneshape,
vector("double", length = 1),
Q = 0.002, h0 = 10, r_well = 0.08, kf = 1*10^-4
)

plot(h_at_x~x_range)

}
