% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiemcore.R
\name{thiem_h}
\alias{thiem_h}
\title{Use the Dupuit-Thiem equation to calculate the water column level in a well}
\usage{
thiem_h(Q, h0, r_well, kf, dh_search_min = 0.001, dh_search_max = h0)
}
\arguments{
\item{Q}{the pumping rate of the well in m³/s}

\item{h0}{the natural height of the water column above the aquitard (in absence of a well) in m}

\item{r_well}{the radius of the well}

\item{kf}{the hydraulic conductivity of the aquifer in m/s}

\item{dh_search_min}{the minimum possible drawdown of the well to be considered in m. Defaults to 0.001m == 1mm}

\item{dh_search_max}{the minimum possible drawdown of the well to be considered in m. Defaults to \code{h0}}
}
\value{
the water column level in the well in m above aquitard
}
\description{
\code{thiem_h} uses a heuristic to determine the height of
the water level above aquitard left in a well that is pumping at rate Q. To
get the often more useful drawdown of a well see \code{\link{thiem_h}}.
}
\examples{
# first example
thiem_h(Q = 0.001, h0 = 10, r_well = 0.08, kf = 10^-4)

}
