% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiemcore.R
\name{thiem_Q}
\alias{thiem_Q}
\title{Use the Dupuit-Thiem equation to calculate the pumping rate of a well}
\usage{
thiem_Q(h1, h2, r1, r2, kf)
}
\arguments{
\item{h1}{the height of the water column above the aquitard at monitoring well 1 in m}

\item{h2}{the height of the water column above the aquitard at monitoring well 2 in m}

\item{r1}{the distance from the pumping well to the monitoring well No 1 in m}

\item{r2}{the distance from the pumping well to the monitoring well No 2 in m}

\item{kf}{the hydraulic conductivity of the aquifer in m/s}
}
\value{
the pumping rate in m³/s
}
\description{
\code{thiem_Q} is an implementation of the widely used
well equation by Dupuit-Thiem FOR UNCONFINED AQUIFERS.
}
\examples{
# first example
thiem_Q(h1 = 10, h2 = 8, r1 = 50, r2 = 0.1, kf = 10^-3)

# second example
drawdown_steps <- seq(0.2, 5, 0.2) # m
aquifer_cond <- 2.5*10^-3 # m/s
monitoring_well_potential = 10 # m water column above aquitard
monitoring_well_distance = 30 # m
pumping_well_radius = 0.08 # m == DN160

rates <- vapply(
X = drawdown_steps,
FUN = function(h2, h1, r1, r2, kf){ # rearranging input order for vapply
  thiem_Q(h1 = h1, h2 = h2, r1 = r1, r2 = r2, kf = kf)
},
FUN.VALUE = vector("double", length = 1),
h1 = monitoring_well_potential,
r1 = monitoring_well_distance,
r2 = pumping_well_radius,
kf = aquifer_cond)

plot(drawdown_steps~rates)

}
